requirements.txt
amqp==2.5.2
appnope==0.1.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
celery==4.4.0
cffi==1.14.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.3
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
vine==1.3.0
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
